Changes to user controller

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 use App\Models\User;
 
 class UserController extends Controller
@@ -10,10 +11,27 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $users = User::all();
-        return response()->json($users);
+        $headerValidator = Validator::make($request->headers->all(), [
+            'client-id' => 'required|max:255',
+        ]);
+        if ($headerValidator->fails()) {
+            return response()->json(['message' => 'Header validation failed',
+                'error' => $headerValidator->errors()->first()], 400);
+        }
+        try {
+            $users = User::where('client-id', $request->header('client-id'))
+                ->get();
+            if ($users) {
+                return response()->json($users);
+            } else {
+                return response()->json(['message' => 'No users found'], 404);
+            }
+        } catch (\Exception $e) {
+            return response()->json(['message' => 'User retrieval failed',
+                'error' => $e->getMessage()], 500);
+        }
     }
 
     /**
@@ -21,12 +39,24 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $headerValidator = Validator::make($request->headers->all(), [
+            'client-id' => 'required|max:255',
+        ]);
+        if ($headerValidator->fails()) {
+            return response()->json(['message' => 'Header validation failed',
+                'error' => $headerValidator->errors()->first()], 400);
+        }
         try {
             $validatedData = $request->validate([
                 'name' => 'required|string|max:255',
-                'email' => 'required|email|max:255',
-                'client_id' => 'required|string|max:255',
+                'email' => 'required|max:255',
             ]);
+        } catch (\Exception $e) {
+            return response()->json(['message' => 'Validation failed',
+                'error' => $e->getMessage()], 400);
+        }
+        try {
+            $validatedData['client-id'] = $request->header('client-id');
             $user = User::create($validatedData);
             return response()->json($user, 201);
         } catch (\Exception $e) {
@@ -40,13 +70,17 @@ class UserController extends Controller
      */
     public function show(Request $request)
     {
+        $headerValidator = Validator::make($request->headers->all(), [
+            'email' => 'required|max:255',
+            'client-id' => 'required|max:255',
+        ]);
+        if ($headerValidator->fails()) {
+            return response()->json(['message' => 'Header validation failed',
+                'error' => $headerValidator->errors()->first()], 400);
+        }
         try {
-            $validatedData = $request->validate([
-                'email' => 'required|email|max:255',
-                'client_id' => 'required|string|max:255',
-            ]);
-            $user = User::where('email', $validatedData['email'])
-                ->where('client_id', $validatedData['client_id'])->first();
+            $user = User::where('email', $request->header('email'))
+                ->where('client-id', $request->header('client-id'))->first();
             if ($user) {
                 return response()->json($user);
             } else {
@@ -63,13 +97,26 @@ class UserController extends Controller
      */
     public function update(Request $request)
     {
+        $headerValidator = Validator::make($request->headers->all(), [
+            'client-id' => 'required|max:255',
+        ]);
+        if ($headerValidator->fails()) {
+            return response()->json(['message' => 'Header validation failed',
+                'error' => $headerValidator->errors()->first()], 400);
+        }
         try {
             $validatedData = $request->validate([
-                'email' => 'required|email|max:255',
+                'email' => 'required|max:255',
                 'name' => 'sometimes|string|max:255',
             ]);
+        } catch (\Exception $e) {
+            return response()->json(['message' => 'Validation failed',
+                'error' => $e->getMessage()], 400);
+        }
+        try {
+            $validatedData['client-id'] = $request->header('client-id');
             $user = User::where('email', $validatedData['email'])
-                ->where('client_id', $validatedData['client_id'])->first();
+                ->where('client-id', $validatedData['client-id'])->first();
             if ($user) {
                 $user->update($validatedData);
                 return response()->json($user);
@@ -87,12 +134,17 @@ class UserController extends Controller
      */
     public function destroy(Request $request)
     {
+        $headerValidator = Validator::make($request->headers->all(), [
+            'email' => 'required|max:255',
+            'client-id' => 'required|max:255',
+        ]);
+        if ($headerValidator->fails()) {
+            return response()->json(['message' => 'Header validation failed',
+                'error' => $headerValidator->errors()->first()], 400);
+        }
         try {
-            $validatedData = $request->validate([
-            'email' => 'required|email|max:255',
-            ]);
-            $user = User::where('email', $validatedData['email'])
-                ->where('client_id', $validatedData['client_id'])->first();
+            $user = User::where('email', $request->header('email'))
+                ->where('client-id', $request->header('client-id'))->first();
             if ($user) {
                 $user->delete();
                 return response()->json(['message' => 'User deleted successfully']);
